std: use SIGWINCH signal to update the size
SIGWINCH is a signal send to the process when the host terminal is
resized, note that it's only available on Unix platforms. Catching this
signal allows us to dynamically adapt the number of columns at a very
low cost compared to `dynamic_ncols`. Indeed, as this approach is signal
based, there is no need to poll the terminal size on each iteration.

Signed-off-by: Lucas CHOLLET <lucas.chollet@free.fr>